Refactor card type categorization logic to DeckCards collection class
